feat:  添加QCloud的实现

+++ myddd-vertx-media/myddd-vertx-media-storage/myddd-vertx-media-storage-local/src/main/kotlin/org/myddd/vertx/media/storeage/LocalMediaStorage.kt
@@ -5,6 +5,7 @@ import io.vertx.core.Vertx
 import io.vertx.core.json.JsonObject
 import io.vertx.kotlin.coroutines.await
 import org.myddd.vertx.base.BusinessLogicException
+import org.myddd.vertx.file.FileDigest
 import org.myddd.vertx.ioc.InstanceFactory
 import org.myddd.vertx.media.domain.MediaErrorCode
 import org.myddd.vertx.media.domain.MediaExtra
@@ -14,20 +15,27 @@ import java.time.LocalDateTime
 
 class LocalMediaStorage(private var storagePath: String = System.getProperty("java.io.tmpdir") + "STORAGE") :MediaStorage {
 
-    companion object {
-        private val randomIdString by lazy { InstanceFactory.getInstance(RandomIDString::class.java) }
-        private val vertx by lazy { InstanceFactory.getInstance(Vertx::class.java) }
-        private val fs = vertx.fileSystem()
-    }
+    private val fileDigest by lazy { InstanceFactory.getInstance(FileDigest::class.java) }
+    private val vertx by lazy { InstanceFactory.getInstance(Vertx::class.java) }
 
     override suspend fun uploadToStorage(tmpPath: String):Future<MediaExtra> {
         return try {
-            val destPath = randomFilePath().await()
+            val fs = vertx.fileSystem()
+
             val exists = fs.exists(tmpPath).await()
             if(!exists){
                 throw BusinessLogicException(MediaErrorCode.SOURCE_FILE_NOT_EXISTS)
             }
-            fs.copy(tmpPath,destPath).await()
+
+            val destDir = randomFilePath().await()
+            val digest = fileDigest.digest(tmpPath).await()
+            val destPath ="$destDir/$digest"
+
+            val destFileExists = fs.exists(destPath).await()
+            if(!destFileExists){
+                fs.copy(tmpPath,destPath).await()
+            }
+
             Future.succeededFuture(LocalMediaExtra(path = destPath))
         }catch (t:Throwable){
             Future.failedFuture(t)
@@ -36,6 +44,7 @@ class LocalMediaStorage(private var storagePath: String = System.getProperty("ja
 
     override suspend fun downloadFromStorage(extra: MediaExtra): Future<String> {
         return try {
+            val fs = vertx.fileSystem()
             val localMediaExtra = extra as LocalMediaExtra
             val exists = fs.exists(localMediaExtra.path).await()
             if(!exists){
@@ -54,14 +63,14 @@ class LocalMediaStorage(private var storagePath: String = System.getProperty("ja
 
     internal suspend fun randomFilePath():Future<String>{
         return try {
+            val fs = vertx.fileSystem()
             val now = LocalDateTime.now()
-            val dir = "$storagePath/${now.year}/${now.monthValue}/${now.dayOfMonth}/${now.hour}"
+            val dir = "$storagePath/${now.year}/${now.monthValue}/${now.dayOfMonth}"
             fs.mkdirs(dir).await()
 
-            Future.succeededFuture("$dir/${randomIdString.randomUUID()}")
+            Future.succeededFuture(dir)
         }catch (t:Throwable){
             Future.failedFuture(t)
         }
-
     }
 }
